Handle camera and simulated image in a dirty way
If the image is 640x480, assume it's the camera.  Set the roi to the tiny area that isn't badly distorted, use adapativeThreshhold, shift the result by the roi so it looks nice in the image.  This makes it impossible to scale the roi up to the full size of the image without rewriting some stuff but it's ok for now.